add support for DISPLAYCONFIG_DEVICE_INFO_GET_ADVANCED_COLOR_INFO_2 first (#6)
This commit addresses issue #4 by adding support for `DISPLAYCONFIG_DEVICE_INFO_GET_ADVANCED_COLOR_INFO_2`. This API became available in Windows 10.0.26100.0 and is not compatible with older versions of Windows. If the new API doesn't work, HDRSwitcher falls back to the older API.

Additionally, this commit introduces a `--debug` argument that allows to check which API is being used by the tool.